+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r180-release...4a2f0ed2fe8b77f43f4428138cf8f5b81187f85b)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r181-release...8bd5702ad87999f73722a3bd89294e2bd9adb904)
 
-- Add a "Shape" option for PW:S
+- Update raid debuffs for Dawn of the Infinite
+- Refactor layouts
+- Implemented import & export for click-castings
+- Change click-castings to character profile on Wrath
+- [snippet] SortPartyBySpec